+++ test/plugin/pom.xml
@@ -126,7 +126,7 @@
                         <sourceDirectory>scenarios/canal-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/cassandra-java-driver-3.x-scenario</sourceDirectory>
                         <sourceDirectory>scenarios/correlation-autotag-scenario</sourceDirectory>
-<!--                        <sourceDirectory>scenarios/customize-scenario</sourceDirectory>-->
+                        <sourceDirectory>scenarios/customize-scenario</sourceDirectory>
 <!--                        <sourceDirectory>scenarios/cxf-scenario</sourceDirectory>-->
 <!--                        <sourceDirectory>scenarios/dbcp-2.x-scenario</sourceDirectory>-->
 <!--                        <sourceDirectory>scenarios/dubbo-2.5.x-scenario</sourceDirectory>-->
